Changed the index page to login
Created a render.js file with two methods. The first one, getInfo, adds a user's info to an array that bhai uploaded in data.js. 
The second method, login, checks to see if the username and password matches info in the aforementioned array.
We're struggling to link the pages together, both with the data in the array and the hyperlinks. 
In the next update, can someone work on linking these together and maybe make a user profile page?